Create some docs
This is quite minimal, because I think most of the information
about the Jupyter command, environment variables, etc. should be
in the overview docs, not here. But it seems like it should have
some API docs.

Should we document any of the constants in the paths module, e.g.
SYSTEM_JUPYTER_PATH, as public API? Likewise, do we consider the API of
the JupyterApp base class public?