ensure Python data_files are found
ask distutils where it would install data files for user=True|False,
and ensure those locations are on various paths.

For user=False, this is generally sys.prefix, but not always (system Python on debian, OS X)

In the cases where this is *not* sys.prefix, sys.prefix is still included.