Use GetUserNameEx to get SID of current user to secure win32 files
Found that the SID obtained when using GetUserName was causing permission
issues because it wasn't the actual SID for the current user.  By changing
the call to use `GetUserNameEx` with a parameter of `NameSamCompatible`,
the SID that was returned did indeed correspond to the current user.  In
addition, analysis of the file using Windows explorer also showed the SID
being resolved to the current user's Windows ID, rather than a raw SID (as
before).

Co-authored-by: snapo <6347922+snapo@users.noreply.github.com>